requirements.txt
Django
gunicorn
pandas
plotly
psycopg
psycopg2-binary
channels
djangorestframework
cryptography
channels-redis